d310650 Misc. fixes
f70c7f3 Add extern "C" for use in C++ builds
22bbbc9 Add remove observer logic when delete resource.
da62cb7 Adding brackets at IPv6 address
0f20d9e Update code to use size_t
956d56a Add Unit tests, fixes, and doc for oc_base64

+++ api/oc_collection.c
@@ -157,7 +157,7 @@ oc_link_add_rel(oc_link_t *link, const char *rel)
 }
 
 oc_collection_t *
-oc_get_collection_by_uri(const char *uri_path, size_t uri_path_len, int device)
+oc_get_collection_by_uri(const char *uri_path, size_t uri_path_len, size_t device)
 {
   while (uri_path[0] == '/') {
     uri_path++;
@@ -182,7 +182,7 @@ oc_get_collection_by_uri(const char *uri_path, size_t uri_path_len, int device)
 }
 
 oc_link_t *
-oc_get_link_by_uri(oc_collection_t *collection, const char *uri_path, int uri_path_len)
+oc_get_link_by_uri(oc_collection_t *collection, const char *uri_path, size_t uri_path_len)
 {
   oc_link_t *link = NULL;
 
@@ -195,7 +195,7 @@ oc_get_link_by_uri(oc_collection_t *collection, const char *uri_path, int uri_pa
     link = oc_list_head(collection->links);
     while (link != NULL) {
       if (link->resource &&
-          (int)oc_string_len(link->resource->uri) == (uri_path_len + 1) &&
+          oc_string_len(link->resource->uri) == (uri_path_len + 1) &&
           strncmp(oc_string(link->resource->uri) + 1, uri_path, uri_path_len) == 0) {
         break;
       }
