+++ port/openthread/ipadapter.c
@@ -78,7 +78,7 @@ free_endpoints(void)
 }
 
 oc_endpoint_t *
-oc_connectivity_get_endpoints(int device)
+oc_connectivity_get_endpoints(size_t device)
 {
   (void)device;
 
@@ -152,7 +152,7 @@ oc_send_buffer(oc_message_t *message)
 }
 
 int
-oc_connectivity_init(int device)
+oc_connectivity_init(size_t device)
 {
   (void)device;
 
@@ -203,7 +203,7 @@ oc_connectivity_init(int device)
 }
 
 void
-oc_connectivity_shutdown(int device)
+oc_connectivity_shutdown(size_t device)
 {
   (void)device;
