+++ security/oc_pstat.c
@@ -92,7 +92,7 @@ dump_pstat_dos(oc_sec_pstat_t *ps)
 #endif /* OC_DEBUG */
 
 static bool
-valid_transition(int device, oc_dostype_t state)
+valid_transition(size_t device, oc_dostype_t state)
 {
   switch (pstat[device].s) {
   case OC_DOS_RESET:
@@ -119,9 +119,9 @@ valid_transition(int device, oc_dostype_t state)
   return true;
 }
 
-static bool oc_pstat_handle_state(oc_sec_pstat_t *ps, int device);
+static bool oc_pstat_handle_state(oc_sec_pstat_t *ps, size_t device);
 static bool
-oc_pstat_handle_state(oc_sec_pstat_t *ps, int device)
+oc_pstat_handle_state(oc_sec_pstat_t *ps, size_t device)
 {
   oc_sec_acl_t *acl = oc_sec_get_acl(device);
   oc_sec_doxm_t *doxm = oc_sec_get_doxm(device);
@@ -334,7 +334,7 @@ pstat_state_error:
 }
 
 oc_sec_pstat_t *
-oc_sec_get_pstat(int device)
+oc_sec_get_pstat(size_t device)
 {
 #ifdef OC_DEBUG
   dump_pstat_dos(&pstat[device]);
@@ -343,20 +343,20 @@ oc_sec_get_pstat(int device)
 }
 
 bool
-oc_sec_is_operational(int device)
+oc_sec_is_operational(size_t device)
 {
   return pstat[device].isop;
 }
 
 void
-oc_sec_pstat_default(int device)
+oc_sec_pstat_default(size_t device)
 {
   pstat[device].s = OC_DOS_RESET;
   oc_pstat_handle_state(&pstat[device], device);
 }
 
 void
-oc_sec_encode_pstat(int device)
+oc_sec_encode_pstat(size_t device)
 {
 #ifdef OC_DEBUG
   dump_pstat_dos(&pstat[device]);
@@ -383,13 +383,13 @@ static oc_event_callback_retval_t
 dump_acl_post_otm(void *data)
 {
   size_t device = (size_t)data;
-  oc_sec_dump_acl((int)device);
-  oc_sec_dump_unique_ids((int)device);
+  oc_sec_dump_acl(device);
+  oc_sec_dump_unique_ids(device);
   return OC_EVENT_DONE;
 }
 
 bool
-oc_sec_decode_pstat(oc_rep_t *rep, bool from_storage, int device)
+oc_sec_decode_pstat(oc_rep_t *rep, bool from_storage, size_t device)
 {
   bool transition_state = false;
   oc_sec_pstat_t ps;
@@ -508,7 +508,7 @@ post_pstat(oc_request_t *request, oc_interface_mask_t interface, void *data)
 {
   (void)interface;
   (void)data;
-  int device = request->resource->device;
+  size_t device = request->resource->device;
   if (oc_sec_decode_pstat(request->request_payload, false, device)) {
     oc_send_response(request, OC_STATUS_CHANGED);
     request->response->response_buffer->response_length = 0;
